+++ examples/ArrayHandleZip.cxx
@@ -15,7 +15,7 @@ void CheckArray(const ArrayHandleType array)
   VTKM_TEST_ASSERT(portal.GetNumberOfValues() == 3,
                    "Permuted portal has wrong size.");
 
-  typedef vtkm::Pair<vtkm::Id, vtkm::Float32> PairType;
+  typedef vtkm::Pair<vtkm::Id, vtkm::Float64> PairType;
 
   VTKM_TEST_ASSERT(test_equal(portal.Get(0), PairType(3, 0.0)),
                    "Zipped array has wrong value.");
@@ -33,7 +33,7 @@ void Test()
   typedef vtkm::cont::ArrayHandle<vtkm::Id> ArrayType1;
   typedef ArrayType1::PortalControl PortalType1;
 
-  typedef vtkm::cont::ArrayHandle<vtkm::Float32> ArrayType2;
+  typedef vtkm::cont::ArrayHandle<vtkm::Float64> ArrayType2;
   typedef ArrayType2::PortalControl PortalType2;
 
   // Create an array of vtkm::Id with values [3, 0, 1]
@@ -53,7 +53,7 @@ void Test()
   portal2.Set(2, 0.2);
 
   // Zip the two arrays together to create an array of
-  // vtkm::Pair<vtkm::Id, vtkm::Float32> with values [(3,0.0), (0,0.1), (1,0.2)]
+  // vtkm::Pair<vtkm::Id, vtkm::Float64> with values [(3,0.0), (0,0.1), (1,0.2)]
   vtkm::cont::ArrayHandleZip<ArrayType1,ArrayType2> zipArray(array1, array2);
   ////
   //// END-EXAMPLE ArrayHandleZip.cxx
